app/requirements.txt
azure-keyvault-secrets==4.6.0
azure-identity==1.12.0
azure-mgmt-web==7.0.0
pytest==7.2.1
pytest-mock==3.10.0